# Conflicts:
#   STUDY/Yi/README.md

+++ backend/src/main/java/com/ssafy/api/response/UsersRes.java
@@ -18,7 +18,7 @@ public class UsersRes {
 
 
     @ApiModelProperty(name="user's uid")
-    int uid;
+    Long uid;
     @ApiModelProperty(name="user's email")
     String email;
     @ApiModelProperty(name="user's nickname")
